requirements.txt
streamlit
pandas
scikit-surprise
gdown
pickle5
matplotlib
